requirements.txt
Django >=4.2,<5.2.5
gunicorn==23.0.0
psycopg2>=2.9.6,<3.0
psycopg-binary
djangorestframework
markdown      
django-filter
paho-mqtt
